feat: migrate collapse component
