--- src-v2/components/collpse/index.tsx
@@ -1,4 +1,0 @@
-import { Collapse, CollapseProps } from '@chakra-ui/react';
-
-export default Collapse;
-export type { CollapseProps };
